--- assets/src/application/entities/ticket/index.ts
@@ -1,2 +1,0 @@
-export { default as isOnSale } from './isOnSale';
-export { default as isPending } from './isPending';
