+++ assets/src/domain/shared/predicates/tickets/filters/nextOnSaleOrPendingOnly.ts
