+++ assets/src/domain/shared/predicates/tickets/filters/percentSoldBelow.ts
