+++ assets/src/domain/shared/predicates/tickets/filters/sortTicketEntitiesList.ts
