+++ assets/src/domain/shared/predicates/tickets/filters/validFiniteQuantity.ts
