+++ assets/src/domain/shared/predicates/tickets/filters/validInfiniteQuantity.ts
