+++ assets/src/domain/shared/predicates/tickets/index.ts
@@ -1,1 +1,3 @@
-export { default as filterTicketEntities } from './filterTicketEntities';
+export { default as filterTicketEntities } from './filters';
+export { default as isOnSale } from './isOnSale';
+export { default as isPending } from './isPending';
