+++ assets/src/domain/shared/predicates/tickets/isOnSale.ts
