+++ assets/src/domain/shared/predicates/tickets/isPending.ts
